Cleanup changes and move testable code to a separate file
