+++ src/main/CMakeLists.txt
@@ -498,6 +498,8 @@ main_sources(COMMON_SRC
     io/ledstrip.h
     io/osd.c
     io/osd.h
+    io/osd_utils.c
+    io/osd_utils.h
     io/osd_canvas.c
     io/osd_canvas.h
     io/osd_common.c
