+++ src/test/unit/CMakeLists.txt
@@ -34,7 +34,7 @@ set_property(SOURCE time_unittest.cc PROPERTY depends "drivers/time.c")
 
 set_property(SOURCE circular_queue_unittest.cc PROPERTY depends "common/circular_queue.c")
 
-set_property(SOURCE osd_unittest.cc PROPERTY depends "io/osd.c" "io/displayport_msp_osd.c" "common/typeconversion.c")
+set_property(SOURCE osd_unittest.cc PROPERTY depends "io/osd_utils.c" "io/displayport_msp_osd.c" "common/typeconversion.c")
 set_property(SOURCE osd_unittest.cc PROPERTY definitions OSD_UNIT_TEST USE_MSP_DISPLAYPORT DISABLE_MSP_BF_COMPAT)
 
 function(unit_test src)
